+++ src/View/Tag.php
@@ -60,38 +60,51 @@ abstract class Tag
      */
     public function arguments($arguments = [])
     {
-        foreach ($arguments as $key => $value) {
-            if ($this->hasArgumentMutator($key)) {
-                $value = $this->setMutatedArgumentValue($key, $value);
-            }
-
-            $this->arguments[$key] = $value;
-        }
+        $this->arguments = $arguments;
 
         return $this;
     }
 
     /**
-     * Determine if a set mutator exists for a argument.
+     * Set the value for an argument
+     *
+     * @param  string  $key
+     * @param  mixed  $value
+     * @return static
+     */
+    public function setArgument($key, $value)
+    {
+        $this->arguments[$key] = $value;
+
+        return $this;
+    }
+
+    public function getArgumentFallback($key, $value)
+    {
+        return $value;
+    }
+
+    /**
+     * Determine if a get mutator exists for a argument.
      *
      * @param  string  $key
      * @return bool
      */
     public function hasArgumentMutator($key)
     {
-        return method_exists($this, 'set'.Str::studly($key).'Argument');
+        return method_exists($this, 'get'.Str::studly($key).'Argument');
     }
 
     /**
-     * Set the value of a argument using its mutator.
+     * Get the value of a argument using its mutator.
      *
      * @param  string  $key
      * @param  mixed  $value
      * @return mixed
      */
-    protected function setMutatedArgumentValue($key, $value)
+    protected function getMutatedArgumentValue($key, $value)
     {
-        return $this->{'set'.Str::studly($key).'Argument'}($value);
+        return $this->{'get'.Str::studly($key).'Argument'}($value);
     }
 
     /**
@@ -114,14 +127,47 @@ abstract class Tag
      */
     public function getArgument(string $key): mixed
     {
+        $value = null;
+
         if ($this->hasArgument($key)) {
-            return $this->arguments[$key];
+            $value = $this->arguments[$key];
+        }
+        // If a default value was defined on the parameter use it
+        elseif ($this->parameters()->has($key)) {
+            $value = $this->parameters()->get($key)->defaultValue ?? null;
         }
 
-        // If a default value was defined on the parameter use it
-        if ($this->parameters()->has($key)) {
-            return $this->parameters()->get($key)->defaultValue ?? null;
+        if ($this->hasArgumentMutator($key)) {
+            $value = $this->getMutatedArgumentValue($key, $value);
+        } else {
+            $value = $this->getArgumentFallback($key, $value);
         }
+
+        return $value;
+    }
+
+    /**
+     * Get a list of all mutated arguments
+     *
+     * @return array
+     */
+    public function getArguments(): array
+    {
+        $defaults = $this->parameters()->filter(function ($parameter) {
+            return ! is_null($parameter->defaultValue);
+        })->transform(function ($parameter) {
+            return $parameter->defaultValue;
+        });
+
+        $arguments = $defaults->merge($this->arguments);
+
+        return $arguments->transform(function ($value, $key) {
+            if ($this->hasArgumentMutator($key)) {
+                return $this->getMutatedArgumentValue($key, $value);
+            } else {
+                return $this->getArgumentFallback($key, $value);
+            }
+        })->toArray();
     }
 
     /**