+++ src/View/TagBuilder.php
@@ -5,7 +5,7 @@ namespace Expressionengine\Coilpack\View;
 use Expressionengine\Coilpack\Facades\Coilpack;
 use Traversable;
 
-class TagBuilder extends Tag implements \IteratorAggregate
+class TagBuilder extends LegacyTag
 {
     /**
      * The name of the module
@@ -13,6 +13,13 @@ class TagBuilder extends Tag implements \IteratorAggregate
      * @var string
      */
     protected $name;
+
+    /**
+     * The addon instance
+     *
+     * @var mixed
+     */
+    protected $instance;
 
     /**
      * The method being called on the module
@@ -88,59 +95,59 @@ class TagBuilder extends Tag implements \IteratorAggregate
     //     return $this->process();
     // }
 
-    public function __toString()
-    {
-        $output = $this->run();
+    // public function __toString()
+    // {
+    //     $output = $this->run();
 
-        // Automatically unwrap an array with single item when converting to string
-        if ($output instanceof \Countable && count($output) === 1) {
-            return (string) $output[0];
-        }
+    //     // Automatically unwrap an array with single item when converting to string
+    //     if ($output instanceof \Countable && count($output) === 1) {
+    //         return (string) $output[0];
+    //     }
 
-        return $output;
-    }
+    //     return $output;
+    // }
 
-    public function run()
-    {
-        $output = Coilpack::isolateTemplateLibrary(function ($template) {
-            $output = $this->getInstanceClass()->{$this->method}();
-            // If the Tag stored data for us in the template library that is preferable to the generated output
-            return $template->get_data() ?: $output;
-        }, $this->arguments);
+    // public function run()
+    // {
+    //     $output = Coilpack::isolateTemplateLibrary(function ($template) {
+    //         $output = $this->getInstanceClass()->{$this->method}();
+    //         // If the Tag stored data for us in the template library that is preferable to the generated output
+    //         return $template->get_data() ?: $output;
+    //     }, $this->arguments);
 
-        return is_array($output) && is_array(current($output)) ? collect($output)->map(function ($row) {
-            return \Expressionengine\Coilpack\TemplateOutput::make()->value($row);
-        }) : \Expressionengine\Coilpack\TemplateOutput::make()->value($output);
-    }
+    //     return is_array($output) && is_array(current($output)) ? collect($output)->map(function ($row) {
+    //         return \Expressionengine\Coilpack\TemplateOutput::make()->value($row);
+    //     }) : \Expressionengine\Coilpack\TemplateOutput::make()->value($output);
+    // }
 
-    private function getInstanceClass()
-    {
-        $class = '\\'.ltrim($this->instance->getFrontendClass(), '\\');
+    // private function getInstanceClass()
+    // {
+    //     $class = '\\'.ltrim($this->instance->getFrontendClass(), '\\');
 
-        return new $class;
-    }
+    //     return new $class;
+    // }
 
-    private function processArguments($value)
-    {
-        if (is_array($value)) {
-            if (count($value) > 1) {
-                return $value;
-            } else {
-                $value = current($value);
-            }
-        }
+    // private function processArguments($value)
+    // {
+    //     if (is_array($value)) {
+    //         if (count($value) > 1) {
+    //             return $value;
+    //         } else {
+    //             $value = current($value);
+    //         }
+    //     }
 
-        return (is_string($value) && strpos($value, ',')) ? explode(',', $value) : $value;
-    }
+    //     return (is_string($value) && strpos($value, ',')) ? explode(',', $value) : $value;
+    // }
 
-    public function getIterator(): Traversable
-    {
-        $output = $this->run();
+    // public function getIterator(): Traversable
+    // {
+    //     $output = $this->run();
 
-        if ($output instanceof \IteratorAggregate) {
-            return $output->getIterator();
-        }
+    //     if ($output instanceof \IteratorAggregate) {
+    //         return $output->getIterator();
+    //     }
 
-        return new \ArrayIterator(is_array($output) ? $output : []);
-    }
+    //     return new \ArrayIterator(is_array($output) ? $output : []);
+    // }
 }